Move scripts and tools into dev directory.
