+++ buildkite.yml
@@ -3,10 +3,19 @@ env:
 
 steps:
   - label: ':white_check_mark:'
-    command: ./dev/check_all.sh
+    command: ./dev/ci/check_all.sh
 
   # - label: ':lipstick: :lint-roller:'
-  #   command: ./dev/lint.sh
+  #   command: ./dev/ci/lint.sh
+
+  - label: ':typescript:'
+    command: cd extensions/go && yarn run build
+
+  - label: ':typescript:'
+    command: cd extensions/template && yarn run build
+
+  - label: ':typescript:'
+    command: cd extensions/typescript && yarn run build
 
   - label: ':jest:'
-    command: ./dev/test.sh
+    command: ./dev/ci/test.sh
