+++ dev/ci/test.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 set -ex
-cd $(dirname "${BASH_SOURCE[0]}")/..
+cd $(dirname "${BASH_SOURCE[0]}")/../..
 
 yarn --mutex network --frozen-lockfile --network-timeout 60000
 yarn run test --verbose
