+++ dev/ci/yarn-deduplicate.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 set -ex
-cd $(dirname "${BASH_SOURCE[0]}")/..
+cd $(dirname "${BASH_SOURCE[0]}")/../..
 
 yarn --mutex network --frozen-lockfile --network-timeout 60000
 yarn run -s yarn-deduplicate --fail --list --strategy fewer ./yarn.lock || {
