+++ dev/scripts/publish.ts
